move session management for requests from server into http module
