+++ utils/http.py
@@ -4,7 +4,15 @@ import time
 import json
 import uuid
 import requests
+import os
+import errno
+import logging
+import traceback
+import pickle
 
+logger = logging.getLogger(__name__)
+
+user_agent = 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:45.0) Gecko/20100101 Firefox/45.0'
 
 def encode_datetime(obj):
     if isinstance(obj, date):
@@ -32,7 +40,7 @@ class FileDownloader:
     def __init__(self):
         self.session = requests.Session()
         self.session.headers.update({
-            'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:45.0) Gecko/20100101 Firefox/45.0'
+            'User-Agent': user_agent
         })
 
     def download_file(self, url, file_path):
@@ -45,3 +53,49 @@ class FileDownloader:
             for chunk in r.iter_content(chunk_size=1024):
                 if chunk:
                     f.write(chunk)
+
+class BangumiRequest:
+
+    def __init__(self):
+
+        # persist request for accessing bangumi api
+        self.session = requests.Session()
+        self.session.headers.update({
+            'User-Agent': user_agent,
+            'Accept': 'application/json;charset=utf-8'
+        })
+
+        session_storage_path = './.session'
+        self.api_bgm_tv_session_path = session_storage_path + '/api_bgm_tv'
+
+        try:
+            if not os.path.exists(session_storage_path):
+                os.makedirs(session_storage_path)
+                logger.info('create session storage dir %s successfully' % session_storage_path)
+        except OSError as exception:
+            if exception.errno == errno.EACCES:
+                raise exception
+            else:
+                logger.warn(exception)
+
+    def __get_cookie_from_storage(self):
+        try:
+            with open(self.api_bgm_tv_session_path, 'r') as f:
+                self.session.cookies = requests.utils.cookiejar_from_dict(pickle.load(f))
+        except Exception as error:
+            logger.warn(traceback.format_exc(error))
+
+    def __save_cookie_to_storage(self):
+        try:
+            with open(self.api_bgm_tv_session_path, 'w') as f:
+                pickle.dump(requests.utils.dict_from_cookiejar(self.session.cookies), f)
+        except Exception as error:
+            logger.warn(traceback.format_exc(error))
+
+    def get(self, url):
+        self.__get_cookie_from_storage()
+        r = self.session.get(url)
+        self.__save_cookie_to_storage()
+        return r
+
+bangumi_request = BangumiRequest()
